Fix NPE crash in back click listener.
Unders some conditions it is possible that back click listener is triggered after the fragment is already unmounted. This is due to the fact that the fragment lifecycle can be controlled from JS. We need to add extra null checks to prevent from crashing in this case. There is no need to handle this case separately as we cannot dismiss screen fragment that is already unmounted.