Use default insets for toolbar icons on Android to match iOS behavior
Previously we were resetting insets on Android while keeping default insets on iOS. This inconsistency was added to make both right and left side insets equal on Android in order to keep symmetry when both right and left icons are added. With this change we adapt default inset settings and apply it to the right end as well to keep symmetry. We also reset navigation start inset (distance between back button and title) in case title is not present. Thanks the latter change we get extra space for custom center subviews.