Fix modal update logic
This change fixes a problem with modal updating logic. First issue we are addressing here is the re-entry problem. We need to avoid re-entry to setModalViewControllers method as it relies on keeping reactSubviews and _presentedModals in sync. Previously we had a re-entry check added but it was a bit too late in the method code. The re-entry check should be added before we access _presentedModals. We've been testing if the list of controllers vs _presentedModals differes at the very beggining. However, if the transition is already running the _presentedModals may be out of sync and that check may fail even though it shouldn't. Placing it after re-entry check guarantees that we will execute it after the syncing process is complete.

The second problem was related to nested stack herarchies with modals. We've been relying on presentedViewController indicating whether a given stack is actually presenting a modal. This however turn out to be misleading as when showing a modal it may actually update presentedViewController for a number of components. As a result we've been getting presentedViewController prop set in a parent VC when presenting from child and when updating modals we'd incorrectly dismiss that modal.