fix: disable gestures in native stack on Android (#496)
If ScreenStacks of native-stack navigators have the same FragmentManager (because each of them is not under a different Screen component), changing a screen in one of the stacks pops one screen in the other stack because of the dismiss() in /android/src/main/java/com/swmansion/rnscreens/ScreenStack.java@master#L34  being called in both ScreenStacks by the FragmentManager. The workaround is to set gestureEnabled to false in createNativeStackNavigator on Android since React-Navigation handles hw back button and setting it will make the mFragmentManager.addOnBackStackChangedListener(mBackStackListener) in /android/src/main/java/com/swmansion/rnscreens/ScreenStack.java@master#L283  not to be called because the screen is not dismissable without gestureEnabled set to true. Should fix #402. PR also fixes some indentations since the file was not indented properly.