feat: enable dev menu on Android (#602)
Overridded method that is called on the key event on which the developer menu on Android should be opened. Change added only for debug mode. Should resolve #276.
Clicking cmd+m on macOS in debug mode should open the developer menu on Android, but Action Bar wants to handle that key event in order to trigger showOverflowMenu on the Toolbar. We then override this method and call React's showDevOptionsDialog in it.