fix: add appContainer to modal screens (#614)
Following the implementation of Modal in react-native: https://github.com/facebook/react-native/blob/master/Libraries/Modal/Modal.js#L206, we attach each screen that is presented modally with an AppContainer that has a listener for LogBox and ElementInspector. This allows for these elements to pop up over the modal screen. There will be 2 ElementInspectors on the first screen due to 2 AppContainer components rendered in it. If we added check for if the index === 0, changing that index would unmount and remount screen content while calling e.g. navigation.reset().