fix: scrollview behavior on transitioning between screens (#594)
We traverse the view in order to find ScrollViews and invoke `scrollViewDidEndDecelerating` on them. It is needed due to the situation where we start scrolling, and before the ScrollView stops decelerating, we move to another screen and as a result, the `scrollViewDidEndDecelerating` is not called on that ScrollView. Due to JS implementation of ScrollView, it causes the elements of the ScrollView not to respond to touches after going back to that screen.