fix: make header subviews subscribe to transition (#820)
Added change to ScreenStackHeaderConfig.java to expose the Toolbar, of which the ScreenStackHeaderSubviews are the children. It is needed since the Toolbar is not a child of Screen in the view hierarchy, therefore its children are not subscribed to the transition by default. Then, in the prepareOutTransition, we add a check for ScreenStackHeaderConfig and apply the transition to the children of its Toolbar making them stay on the screen for during the transition of going back.