fix: blurView integration with ScreenStack (#1406)
BlurView sometimes calls draw method of the app's rootView, causing unexpected behavior in our library. Since BlurView is a child of ScreenStack in the crashing example, calling op.draw() calls BlurView onDraw method, which then sometimes calls draw method in app's rootView. It results in visiting the drawAndRelease method again. The important thing is drawingOps.indices makes IntRange under the hood, which is then used in next loop calls. At the same time, after the most nested for loop ends, it clears the drawingOps, but this change is not visible in the former for since it just loops over IntRange created before. This results in indexOutOfBounds crash when we try to obtain an op from empty drawingOps. As an addition, If we didn't use indices, but for (op in drawingOps), it would result in ConcurrentModificationException.

Current fix makes a copy of drawingOps which is then used in for loop, and clearing the drawingOps so the nested calls cannot change the original one.