fix: enforce non-local visibility of getDynamic symbol of libreact_codegen_rnscreens.so library
This symbol for some, unknown yet reason, is not visible during linking
of lib<project-name>.so (exactly during linking rncli.cpp.o file). We
enforce desired visibility with compiler attributte. Note however that
this solution should be only temporary as the exact reason should be
tracked down and the issue shall be fixed with more generic approach.