fix: maintain native behavior of iOS back button with custom style
Previously if a custom font or font size was set in the config, a new
`RNSUIBarButtonItem` would be created and set as the back bar button item.
However, this would break native behavior where the button title would shorten
to "Back" if there's not enough space. In iOS 13 it's possible to keep this
native behavior and also style the back button using `UINavigationBarAppearance`.

The only caveat with this approach is that because `UINavigationBarAppearance` is
not available prior to iOS 13, so on older systems we will go ahead and create
a custom `RNSUIBarButtonItem`.

This solves some issues mentioned in #1589.