chore: clarify installation instructions for Android (#1633)
Giving more detail to where the override should be.

## Description


Description and motivation for this PR.

Helps users who are experiencing crashing from react-native-screens on
Android who have placed the override in the wrong location. This
modifies the readme for more clear instructions on what to do. When
placed in the MainActivityDelegate, the crash still occurs, while
placing it in the parent, it seems to be fine.

Fixes #1481 .


## Test code and steps to reproduce
N/A

## Checklist
N/A

Co-authored-by: Kacper Kafara <kacperkafara@gmail.com>