Ensure full typings generation from index.native.tsx
The `index.native.d.ts` file that is generated during build doesn't contain
all exported symbols because the source file doesn't use proper `export`
statements, but instead assigns to the CommonJS `module.exports` variable.

This is a problem when consuming the package with tools that support
suffix-based module resolution like the `moduleSuffixes` setting of `tsc`.
The tool resolves the `index` to `index.native` and tries to use the
corresponding declaration file, which is incomplete and the compilation fails.

Note: this change doesn't address the problem that assigning to
`module.exports` is incompatible with ES6 modules and therefore the "module"
build target is unusable in that context.