fix(iOS): image loading for back button on Fabric (#1688)
## Description

Up to now there was missing implementation of back button image loading
logic on Fabric.
I managed to implement it using only RN API (I avoided creation of
custom image component with custom state as it is done in
`react-native-svg`).

It is worth to notice that there might be still some buggy behaviour
present, as instead of provided asset (an arrow, see examples) a red
rectangle is rendered. However I decided to not include further fix in
this PR, as current behaviour is the same on both architectures. Most
likely there will be a follow up PR.

## Changes

* Implemented missing logic
* Added some needed categories (see changes)

## Test code and steps to reproduce

See `Test550` in `FabricTestExample`. 

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes (failure of iOS e2e is a known issue)