fix(iOS): buggy search bar / large title behaviour on Fabric (#1825)
## Description

`UIKit` requires `ScrollView` to be at index 0 in given view's subview
array to enable it's interaction with navigation bar. On Fabric our
`MaybeNestedStack` view got flattened, however it was not removed from
hierarchy -- instead the view was attached as first child of the
`RNSScreenView` disabling system interaction between navigation bar and
a scroll view.

## Changes

1. Added `collapsable={false}` to `MaybeNestedStack` view, preventing it
from being flattened by React Native.
2. Moved `HeaderConfig` component "down" -- so it is attached as second
child of a `Screen`. It was necessary, because currently on Fabric we
are adding `RNSScreenStackHeaderConfig` into subviews of the `Screen`
and it also interrupted the interaction.
3. Refactored code that relied on `RNSScreenStackHeaderConfig` being
first child of a `Screen` (if present at all).

Initially I've played around with fixing (modyfing) the view hierarchy
on the native side just by removing / adding some views when scroll view
was being attached under the `RNSScreenView`, however such approach led
to cascade of bugs (fixing first, caused another one and so on). We can
assume that in general discrepancy between native & shadow tree leads to
unpleasant bugs.

Possible alternative solution would be to not add
`RNSScreenStackHeaderConfig` as a subview of `RNSScreenView` on Fabric.
I did not research it, but it seems that it would require more changes.

**Note**

We need to consult with `@react-navigation` whether the
`@react-navigation/native-stack` has to be adjusted.

### Before


https://github.com/software-mansion/react-native-screens/assets/50801299/3b89dee8-a307-44ba-a9f4-02e4dfdd95ac


### After


https://github.com/software-mansion/react-native-screens/assets/50801299/205f24d4-590b-402e-8b0a-848fd4482d59


## Test code and steps to reproduce

`Test1097` in `FabricTestExample` can be used to test these changes.
Just use the scroll view and observe that the navigation bar gets
stretched / dwindled while scrolling down / up respectively.

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes