Remove unnecessary calls of `calculateHeaderHeightIsModal` on modal dismisses
Finally, a long-debated topic has met its end. I've managed to dive deeper how NativeStackView is currently storing values and reduces the amount of calls to `calculateHeaderHeightIsModal` on iOS. In case when bugs related to the `use(Re)animatedHeaderHeight` hook will occur in the future, we should consider looking into this commit in case when there will be a need to increase amount of calls to this method.
This change has been tested on ScreensExample/TestsExample iOS environment, including nested stacks.