fix: ensure status bar & orientation are updated on container update
Status bar, Navigation bar and Orientation are changed in the trySetWindowTrait function, called in ScreenFragment that ScreenStackFragment extends - so do not forget the super to apply it correctly!

This ensure status bar, navigation bar and orientation are set, even when Fragment or Activity are not available yet when the props is set, which can happen for example on the first setting of the Props.

A similar fix is also done in the ScreenStackHeaderConfig onUpdate function.