Move DelayedFreeze setImmediate into an effect
Executing side effects in render is usually considered bad manners in
react land, and given concurrency, could have unexpected results if
renders are thrown away.  This change makes it so setImmediate fires in
an effect, and also cleans up after itself if for whatever reason the
callback isn't fired. (although I think it's probably not possible with
how setImmediate is implemented)