fix: Update React Native to 0.73.4, change Cocoapods version (#2032)
## Description

Because of the bug with Cocoapods, our CIs are failing on the cached
dependencies. This issue has been fixed in React Native 0.73.3 already,
but this version requires us to update Cocoapods version. This PR
updates our example projects to 0.73.4 version (which is the newest at
the moment) and updates the version of Cocoapods.
You can read more about the problem
[here](https://github.com/facebook/react-native/issues/42698).

## Changes

- Updated examples that are matching 0.73 version to 0.73.4
- Updated Cocoapods version pattern in Gemfiles

## Checklist

- [x] Ensured that CI passes