chore: configure yarn version in package.json (#2077)
## Description

There's no explicit mention or configuration in this repository of which
package manager version is used. Because of that contributors have to
guess what version of yarn they're supposed to use, so they can prevent
unnecessary changes to `yarn.lock`. It looks like this is Yarn 1.22.19,
so this should be added to `package.json`.

## Changes

- Updated `package.json` with `packageManager` set to `yarn@1.22.22`.