fix(iOS): Fix screen traversing on Fabric, update React Native to rc.6 (#2090)
## Description

Some time ago, I've introduced a
[fix](https://github.com/software-mansion/react-native-screens/commit/3b35894f1f83bc4c9daa518d68f3bcb0e832129a#diff-547f3e256039a77611e8a361cd680f6c4ddd17f1e057dea113777001a7b37cc3R617)
for Fabric for scenario, when user tries to navigate to two (or more)
different screens at the same time. Unfortunately, regression shows that
it's not possible for now to navigate on more than one screen, because
of this change. This PR reverts that change.

## Changes

- Reverted a fix for navigating to more than one screen at the same time
- Updated React Native in FabricTestExample to 0.74.0-rc.6

## Screenshots / GIFs

### Before


https://github.com/software-mansion/react-native-screens/assets/23281839/779d1b43-56ae-4a55-8fc5-0594093ce884

### After


https://github.com/software-mansion/react-native-screens/assets/23281839/ea9e421d-0fad-457c-a374-9fec5aba9f53

## Test code and steps to reproduce

You can use `Test1726` in FabricTestExample to reproduce this scenario
and test that change.

## Checklist

- [ ] Ensured that CI passes