feat: Migrate example apps to use react-navigation (#2084)
## Description

This PR migrates our example apps to use react-navigation. Because we're
right now using react-navigation as a submodule (and we intent to remove
native-stack v5 from codebase), it's mandatory to migrate all of the
example apps to @react-navigation/native-stack. To do this, I've moved
all imports for creating navigator to use native-stack v6 and changed
the naming of the props. Right now all of the examples should work, but
it's important to build `lib` directories from submodule (using `yarn
build`) beforehand.

## Changes

- Moved all navigators to use `@react-navigation/native-stack`
- Changed naming of the props to match native-stack v6
- Changed E2E tests a little

## Test code and steps to reproduce
You can just run example apps to test this change.

## Checklist

- [x] Ensured that CI passes