fix(Fabric): back gesture activates `Pressable` elements (#2131)
## Description

When cancelling touches we previously relied on `touchHandler` instance
being exposed by some react managed view above us (screen stack) in the
view hierarchy (on Paper it was `RCTRootContentView`). However this
changed with some recent Fabric version and no view exposes such
property any longer, despite the fact that touch handler is obviously
still utilised by these views even on Fabric, but the field has been
removed from public API.

Currently on Fabric (as of version `0.74.1`) `RCTSurfaceTouchHandler` is
owned by instance of `RCTFabricSurface` owned by instance of
`RCTSurfaceView` which we can expect to live above us in the view
hierarchy (RN mounts it in the very beginning of application runtime).

To get access to the private touchHandler (which is a
`UIGestureRecognizer`) we observe that it is attached to the
`RCTSurfaceView` and thus it is retained in its `gestureRecognizer`
array (native API).

> [!tip]
> We could potentially cache the instance of `RCTSurfaceView`, but for
now I've come to a conclusion that this is an minor minor computation
(there is only one gesture recognizer in the array & the parent lookup
is O(log n) (I know this is not a balanced tree, but you know what I
mean). However this is always an option for future.

> [!note]
> I thought of adding warning in case of `RCTSurfaceView` being not
present above screen stack in view hierarchy but this might be the case
with `modal` stack presentation.

Closes #2118

## Changes

* Added utility extensions on `RCTTouchHandler` &
`RCTSurfaceTouchHandler` classes to make the code more expressive
* Splited `RNSScreenStackView.cancelTouchesInParent` implementation into
two separate for Paper & Fabric
* In case of Fabric added lookup for `RCTSurfaceView` & got access to
its touch handler.


## Test code and steps to reproduce

`Test2118`

See #2118 for issue description.

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes