chore(Android): fix lint warning of CustomToolbar component (#2176)
## Description

Just noticed while working on #2169 that we got a warning in
CustomToolbar.

Edit: followed review suggestions and suppressed lints for all our view,
where this was requried.

## Changes

Suppressed lint on missing constructors. 

We're safe to miss these there, as this view is constructed only
programatically (we do not inflate any of our views).

## Test code and steps to reproduce

N/A

## Checklist

- [ ] Ensured that CI passes