chore: make eps parameter const & functions constexpr in RectUtil (#2195)
## Description

The `eps` parameter in `compareFrameSizes` should be `const`. 

Both functions can be `constexpr` so I've marked them as such.

> [!note]
When doing a bit of research for this PR I've learned that [`constexpr`
implies `inline` linkage (so also bending the ODR
rule)](https://stackoverflow.com/a/14391320/8635522) for functions (and
it seems that since C++17 also for namespace scoped variables) and must
admit I was not aware of this before, thus I've come to conclusion that
it will be ok to leave `inline` keyword there, so that the intent of
`inline` linkage is clear.


## Test code and steps to reproduce

Build Android / iOS

## Checklist

- [ ] Ensured that CI passes