fix(Android): jumping content on fabric (#2313)
## Description

This PR applies modifications to a previous fix:
https://github.com/software-mansion/react-native-screens/pull/2169 for
fabric only, which has stopped working since RN `0.75`.

In RN `0.74` the `adopt` in `RNSScreenComponentDescriptior.h` was once
called without `stateData` but with children and we could then check if
the `ScreenStackHeaderConfig` is present among them and make adjustments
based on it.

When working on
https://github.com/software-mansion/react-native-screens/pull/2292 it
became clear that the fix does not work anymore. Now the `adopt` is
called either with no children and no `stateData` or with both.
The solution is to move the code to `appendChild` in
`RNSScreenShadowNode.cpp` so we can perform the adjustments as soon as
the children append.

## Changes

- moved code from `adopt` in `RNSScreenComponentDescriptior.h` to newly
added `appendChild` override in `RNSScreenShadowNode.cpp`

## Screenshots / GIFs

### Before


https://github.com/user-attachments/assets/6b76864b-58bb-4c6e-9f5b-a01bb0c88d2a

### After


https://github.com/user-attachments/assets/98931e77-3877-4f67-8b28-f49d2e0f42ff


## Test code and steps to reproduce

- Use `TestHeader` to test this change

## Checklist

- [ ] Included code example that can be used to test this change
- [ ] Updated TS types
- [ ] Updated documentation: <!-- For adding new props to native-stack
-->
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/guides/GUIDE_FOR_LIBRARY_AUTHORS.md
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/native-stack/README.md
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/src/types.tsx
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/src/native-stack/types.tsx
- [ ] Ensured that CI passes

---------

Co-authored-by: Kacper Kafara <kacperkafara@gmail.com>