fix: `HeaderConfig` impacts layout of `Screen`'s contents (#2395)
## Description

This PR fixes headerConfig's incorrect layout with custom subviews after
recent changes in
https://github.com/software-mansion/react-native-screens/pull/2325.

> [!note]
@kkafar: 
Previously, before #2325, all children of the headerConfig component
have been positioned absolutely, thus the headerConfig was always of
height 0, not impacting layout of other components. After #2325,
headerConfig's children are positioned using flexbox. This implies that
it has no longer height of 0, thus it impacts the layout of other other
elements, in particular `ScreenContentWrapper`, which is offset by the
height of the highest header config subview.
>
> The initial idea to solve this was to set `height: 0; overflow:
visible`, however, for some yet unknown reason the subviews become
invisible with such styles set of headerConfig. Note that if you set the
`height: 1` it works as expected.
>
> Due to above hindrance we decided to position the headerConfig
approximately at the position of native header, by setting `top: -100%`.
To prevent the headerConfig from blocking gestures we set
`pointerEvents: 'box-none'`.
>
> In the end I want to note, that it would be best if we came out with
solution that excludes headerConfig from layout as it was before #2325.

## Changes

- added `Test2395.tsx` repro
- adjusted headerConfig's styles


## Screenshots / GIFs

### Before
![Screenshot 2024-10-09 at 12 48
32](https://github.com/user-attachments/assets/532fbce1-27cf-4eeb-9b56-1eb3a9e9fd6f)


### After
![Screenshot 2024-10-09 at 12 47
57](https://github.com/user-attachments/assets/a45b6677-9663-4145-8a0a-ff9bbdf22f89)


## Test code and steps to reproduce

- use `Test2395.tsx` repro

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes

---------

Co-authored-by: Kacper Kafara <kacperkafara@gmail.com>