fix(Android,Fabric): invalid behaviour of `fitToContents` sheet detent (#2398)
## Description

Fixes invalid behaviour of `fitToContents` sheet detent on Fabric. 

The erroneous behaviour has been caused by different order of updates on
Fabric relative to Paper.
On Fabric `onCreateView` (the place where bottom sheet behaviour is
created) is called much later,
Because of this the layoutCallback from `ScreenContentWrapper` fails to
set proper sheet height,
as the `sheetBehaviour` is still `null`.

## Changes

`Screen` now keeps weak reference to `ScreenContentWrapper` and queries
it when needed during behaviour configuration.


## Test code and steps to reproduce

`Test1649`, chante `sheetInitialDetents` to `fitToContents` and see that
it now works on Fabric.

## Checklist

- [x] Included code example that can be used to test this change
- [-] Ensured that CI passes