fix(iOS): `extraLight` blur not working (#2338)
Blur style `UIBlurEffectStyleExtraLight` equals `0` so using it in `if
(config.blurEffect)` makes it that the code for applying
`backgroundEffect` doesn't execute.

With this fix the navigation bar can no longer be fully transparent
because blur effect is always set as `extraLight` by default. So it's up
for a discussion if it should stay that way or if we should make this
value optional.

Fixes #2320.

Added `Test2320` to tests.

- [x] Included code example that can be used to test this change
- [ ] Ensured that CI passes

---------

Co-authored-by: Kacper Kafara <kacperkafara@gmail.com>
(cherry picked from commit fcb169324510c58bd1c6fea98d7bb0ab735adf57)