chore: add support for 0.76-rc.6 (#2407)
This PR bumps example apps to 0.76.0-rc.6.
Removed outdated patches for react-native-reanimated and
react-native-gesture-handler.
Add patch for react-native-safe-area-context.
Reanimated is provided as `.tgz` it should be released with the
necessary fixes any day now.
Tested also on 0.75.4 both paper and fabric seems to work fine.

To run it go to `Example` / `FabricExample`:

```
yarn
npx patch-package

```