refactor!: drop support for native-stack v5 (#2373)
## Description

~This PR intents to remove native-stack v5 implementation from the
repository.~

Plan: 

1. [x] ~Remove public `/native-stack` directory~
2. [ ] ~Remove implementation in `src/native-stack/` directory~
3. [x] ~Remove all references in docs to `native-stack@v5`~
4. [ ] ~Remove all in-code references to `native-stack@v5`~

> [!caution]
~2 & 4 require some more work to be done, because screen transitions API
code (integration with reanimated) makes use of API exposed by
`native-stack@v5`.~
>
> ~I need to figure out how to decouple these first.~

The plan has changed. We now intend to mark native-stack@v5 and
deprecated and destined for removal with one of upcoming **minor**
versions of the library after initial 4.0.0 release. For the time before
removal native-stack@v5 will be shipped with few minor versions of
screens v4 line, however it won't be supported and should be considered
broken (and it is, if you're using native-header).

> [!important]
This change breaks `TestScreenAnimation` as this test requires screen
transition API which is not yet ported to v7.

## Changes

Mark types & factory functions of native stack v5 as deprecated and add
comments explaining our approach.

## Test code and steps to reproduce

N/A

## Checklist

- [x] Ensured that CI passes

---------

Co-authored-by: Maciej Stosio <maciekstosio@users.noreply.github.com>