feat: add a ScreenStackItem component (#2433)
Currently there is a lot of logic to make `Screens` work properly in
native-stack, including:

- Making sure screens is enabled and mark it for native-stack
- Making sure screens knows that it contains a large title
- Rendering a `DebugContainer` for `LogBox`
- Making sure the `DebugContainer` contains appropriate logic and
wrapper
- Rendering a nested stack to be able to show header in modals

This diff consolidates this logic to a new component `ScreenStackItem`.
This makes it possible to move the above logic to the
`react-native-screens` package by moving this component. Moving the
logic will make it easier to fix bugs in screens.

Extracted from
https://github.com/react-navigation/react-navigation/commit/fbc8635d22acadd2487f6bfc94338cbdac0d5a60

---------

Co-authored-by: Kacper Kafara <kacperkafara@gmail.com>