fix(iOS,Paper): fix broken modal go-back animation (#2496)
## Description

We recently fixed left-out modals dismissal when reloading react-native.
For some reason
calling in `dismissViewControllerAnimated:completion:` on **nested**
UINavigationController
(the one responsible for displaying the navigation bar in modal, not the
one reponsible for presentation)
during invalidation messes up with the animation. 

I'm not really sure why this is the case. Earlier we were calling
dismiss on all presented modals,
but not on UINavigationController itself and it worked like a charm. 

I've found out that simply animating the change solves the situation -
we fix the animation and keep nice
modal dismissal on reload.

Fixes #2488

## Changes

:point_up:

## Test code and steps to reproduce

`TestModalNavigation`

## Checklist

- [ ] Included code example that can be used to test this change
- [ ] Updated TS types
- [ ] Updated documentation: <!-- For adding new props to native-stack
-->
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/guides/GUIDE_FOR_LIBRARY_AUTHORS.md
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/native-stack/README.md
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/src/types.tsx
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/src/native-stack/types.tsx
- [ ] Ensured that CI passes