fix: make unstable_sheetFooter usable with react-navigation v7 (#2504)
## Description

Associated PR:

* https://github.com/react-navigation/react-navigation/pull/12258

The footer is currently broken, as in the effect of recent refactors it
erroneously ended up being rendered as child as `ScreenStackItem`, thus
due to internal implementation it is rendered under
`ScreenContentWrapper`. However, it has to be rendered as direct child
of `Screen` component (the same as `HeaderConfig`).
 
Under usual circumstances this would be a breaking change, however the
prop is marked as `unstable`, therefore I'm proceeding.

## Changes

`unstable_sheetFooter` is now rendered by `ScreenStackItem`.

## Test code and steps to reproduce

`Test1649`

## Checklist

- [ ] Ensured that CI passes